Add files via upload
Bank Teller exercise re the Analyze Financial Data with Python course but am experiencing issues with the code from number 20, where the balance for the "checking" account should be 35 (30 + 5), not 5 as shown in my code. Re 21, balance of the "savings" account should now be 5 (10 - 5), not 10 as shown in my code.